Refactor form review page to separate SiP code (#7070)
