+++ test/common/schemaform/review/ReviewPage.unit.spec.jsx
@@ -148,242 +148,6 @@ describe('Schemaform review: ReviewPage', () => {
 
     expect(router.push.calledWith('previous-page'));
   });
-  it('should display SaveLink and SaveStatus for SiP enabled forms if user is logged in', () => {
-    const setData = sinon.spy();
-    const onSubmit = sinon.spy();
-    const router = {
-      push: sinon.spy()
-    };
-    const route = {
-      path: 'testPage',
-      pageList: [
-        {
-          path: 'previous-page'
-        },
-        {
-          path: 'testing',
-          pageKey: 'testPage'
-        },
-        {
-          path: 'next-page'
-        }
-      ],
-      formConfig: {
-        chapters: {
-          chapter1: {
-            pages: {
-              page1: {
-                schema: {}
-              }
-            }
-          },
-          chapter2: {
-            pages: {
-              page2: {
-              }
-            }
-          }
-        }
-      }
-    };
-    const form = {
-      disableSave: false,
-      submission: {
-        hasAttemptedSubmit: false
-      },
-      page1: {
-        schema: {},
-      },
-      page2: {
-        schema: {},
-      },
-      data: {
-        privacyAgreementAccepted: true
-      }
-    };
-
-    const user = {
-      profile: {
-        savedForms: []
-      },
-      login: {
-        currentlyLoggedIn: true
-      }
-    };
-
-    const tree = SkinDeep.shallowRender(
-      <ReviewPage
-        router={router}
-        setData={setData}
-        form={form}
-        user={user}
-        onSubmit={onSubmit}
-        setEditMode={f => f}
-        setPrivacyAgreement={f => f}
-        route={route}
-        location={location}/>
-    );
-
-    expect(tree.everySubTree('SaveStatus')).not.to.be.empty;
-    expect(tree.everySubTree('SaveFormLink')).not.to.be.empty;
-  });
-  it('should not display SaveLink or SaveStatus for non-SiP enabled forms', () => {
-    const setData = sinon.spy();
-    const onSubmit = sinon.spy();
-    const router = {
-      push: sinon.spy()
-    };
-    const route = {
-      path: 'testPage',
-      pageList: [
-        {
-          path: 'previous-page'
-        },
-        {
-          path: 'testing',
-          pageKey: 'testPage'
-        },
-        {
-          path: 'next-page'
-        }
-      ],
-      formConfig: {
-        chapters: {
-          chapter1: {
-            pages: {
-              page1: {
-                schema: {}
-              }
-            }
-          },
-          chapter2: {
-            pages: {
-              page2: {
-              }
-            }
-          }
-        }
-      }
-    };
-    const form = {
-      disableSave: true,
-      submission: {
-        hasAttemptedSubmit: false
-      },
-      page1: {
-        schema: {},
-      },
-      page2: {
-        schema: {},
-      },
-      data: {
-        privacyAgreementAccepted: true
-      }
-    };
-
-    const user = {
-      profile: {
-        savedForms: []
-      },
-      login: {
-        currentlyLoggedIn: true
-      }
-    };
-
-    const tree = SkinDeep.shallowRender(
-      <ReviewPage
-        router={router}
-        setData={setData}
-        form={form}
-        user={user}
-        onSubmit={onSubmit}
-        setEditMode={f => f}
-        setPrivacyAgreement={f => f}
-        route={route}
-        location={location}/>
-    );
-
-    expect(tree.everySubTree('SaveStatus')).to.be.empty;
-    expect(tree.everySubTree('SaveFormLink')).to.be.empty;
-  });
-  it('should not display SaveStatus if user is not logged in', () => {
-    const setData = sinon.spy();
-    const onSubmit = sinon.spy();
-    const router = {
-      push: sinon.spy()
-    };
-    const route = {
-      path: 'testPage',
-      pageList: [
-        {
-          path: 'previous-page'
-        },
-        {
-          path: 'testing',
-          pageKey: 'testPage'
-        },
-        {
-          path: 'next-page'
-        }
-      ],
-      formConfig: {
-        chapters: {
-          chapter1: {
-            pages: {
-              page1: {
-                schema: {}
-              }
-            }
-          },
-          chapter2: {
-            pages: {
-              page2: {
-              }
-            }
-          }
-        }
-      }
-    };
-    const form = {
-      disableSave: true,
-      submission: {
-        hasAttemptedSubmit: false
-      },
-      page1: {
-        schema: {},
-      },
-      page2: {
-        schema: {},
-      },
-      data: {
-        privacyAgreementAccepted: true
-      }
-    };
-
-    const user = {
-      profile: {
-        savedForms: []
-      },
-      login: {
-        currentlyLoggedIn: false
-      }
-    };
-
-    const tree = SkinDeep.shallowRender(
-      <ReviewPage
-        router={router}
-        setData={setData}
-        form={form}
-        user={user}
-        onSubmit={onSubmit}
-        setEditMode={f => f}
-        setPrivacyAgreement={f => f}
-        route={route}
-        location={location}/>
-    );
-
-    expect(tree.everySubTree('SaveStatus')).to.be.empty;
-  });
   it('should submit when valid', () => {
     const formConfig = {
       chapters: {
@@ -742,68 +506,5 @@ describe('Schemaform review: ReviewPage', () => {
     tree.getMountedInstance().handleEdit('testPage', true);
     expect(tree.getMountedInstance().state.viewedPages.has('testPage')).to.be.true;
     expect(setEditMode.calledWith('testPage', true, null));
-  });
-  it('should auto save after change', () => {
-    const formConfig = {
-      chapters: {
-        chapter1: {
-          pages: {
-            page1: {}
-          }
-        },
-        chapter2: {
-          pages: {
-            page2: {}
-          }
-        }
-      }
-    };
-
-    const pageList = [
-      {
-        path: 'previous-page'
-      },
-      {
-        path: 'next-page'
-      }
-    ];
-
-    const form = {
-      submission: {
-        hasAttemptedSubmit: false
-      },
-      data: {
-        privacyAgreementAccepted: false
-      }
-    };
-
-    const user = {
-      profile: {
-        savedForms: []
-      },
-      login: {
-        currentlyLoggedIn: true
-      }
-    };
-
-    const setData = sinon.spy();
-    const autoSave = sinon.spy();
-
-    const tree = SkinDeep.shallowRender(
-      <ReviewPage
-        form={form}
-        user={user}
-        setData={setData}
-        route={{ formConfig, pageList }}
-        setPrivacyAgreement={f => f}
-        location={location}/>
-    );
-
-    const instance = tree.getMountedInstance();
-    instance.debouncedAutoSave = autoSave;
-    instance.setData({});
-
-    expect(setData.called).to.be.true;
-    expect(autoSave.called).to.be.true;
   });
 });
